added xls/main.rb and xls/sub.rb
